+++ dolby_vision/src/lib.rs
@@ -1,8 +1,14 @@
+/// Dolby Vision RPU (as found in HEVC type 62 NALUs) module
 pub mod rpu;
+
+/// SMPTE ST2094-10 metadata module
 pub mod st2094_10;
+
+/// Various utils
 /// cbindgen:ignore
 pub mod utils;
 
+/// Dolby Vision XML metadata module
 #[cfg(feature = "xml")]
 pub mod xml;
 
@@ -10,5 +16,6 @@ pub mod xml;
 #[cfg(any(cargo_c, feature = "capi"))]
 pub mod capi;
 
+/// Structs used and exposed in the C API
 #[cfg(any(cargo_c, feature = "capi"))]
 pub mod c_structs;