+++ dolby_vision/src/st2094_10/metadata_blocks/level2.rs
@@ -3,7 +3,7 @@ use bitvec_helpers::{bitvec_reader::BitVecReader, bitvec_writer::BitVecWriter};
 #[cfg(feature = "serde_feature")]
 use serde::Serialize;
 
-use super::ExtMetadataBlock;
+use super::{ExtMetadataBlock, ExtMetadataBlockInfo};
 
 ///  Creative intent trim passes per target display peak brightness
 #[repr(C)]
@@ -40,5 +40,23 @@ impl ExtMetadataBlockLevel2 {
         writer.write_n(&self.trim_chroma_weight.to_be_bytes(), 12);
         writer.write_n(&self.trim_saturation_gain.to_be_bytes(), 12);
         writer.write_n(&self.ms_weight.to_be_bytes(), 13);
+    }
+}
+
+impl ExtMetadataBlockInfo for ExtMetadataBlockLevel2 {
+    fn level(&self) -> u8 {
+        2
+    }
+
+    fn bytes_size(&self) -> u64 {
+        11
+    }
+
+    fn required_bits(&self) -> u64 {
+        85
+    }
+
+    fn sort_key(&self) -> (u8, u16) {
+        (self.level(), self.target_max_pq)
     }
 }