+++ dolby_vision/src/st2094_10/metadata_blocks/level4.rs
@@ -3,8 +3,9 @@ use bitvec_helpers::{bitvec_reader::BitVecReader, bitvec_writer::BitVecWriter};
 #[cfg(feature = "serde_feature")]
 use serde::Serialize;
 
-use super::ExtMetadataBlock;
+use super::{ExtMetadataBlock, ExtMetadataBlockInfo};
 
+/// Something about temporal stability
 #[repr(C)]
 #[derive(Debug, Default, Clone)]
 #[cfg_attr(feature = "serde_feature", derive(Serialize))]
@@ -24,5 +25,19 @@ impl ExtMetadataBlockLevel4 {
     pub fn write(&self, writer: &mut BitVecWriter) {
         writer.write_n(&self.anchor_pq.to_be_bytes(), 12);
         writer.write_n(&self.anchor_power.to_be_bytes(), 12);
+    }
+}
+
+impl ExtMetadataBlockInfo for ExtMetadataBlockLevel4 {
+    fn level(&self) -> u8 {
+        4
+    }
+
+    fn bytes_size(&self) -> u64 {
+        3
+    }
+
+    fn required_bits(&self) -> u64 {
+        24
     }
 }