+++ dolby_vision/src/st2094_10/metadata_blocks/level6.rs
@@ -5,7 +5,7 @@ use serde::Serialize;
 
 use crate::st2094_10::generate::Level6Metadata;
 
-use super::ExtMetadataBlock;
+use super::{ExtMetadataBlock, ExtMetadataBlockInfo};
 
 /// ST2086 metadata fallback
 #[repr(C)]
@@ -40,5 +40,19 @@ impl ExtMetadataBlockLevel6 {
         self.min_display_mastering_luminance = meta.min_display_mastering_luminance;
         self.max_content_light_level = meta.max_content_light_level;
         self.max_frame_average_light_level = meta.max_frame_average_light_level;
+    }
+}
+
+impl ExtMetadataBlockInfo for ExtMetadataBlockLevel6 {
+    fn level(&self) -> u8 {
+        6
+    }
+
+    fn bytes_size(&self) -> u64 {
+        8
+    }
+
+    fn required_bits(&self) -> u64 {
+        64
     }
 }