+++ dolby_vision/src/xml/mod.rs
@@ -1,3 +1,4 @@
+/// XML metadata parser
 mod parser;
 
 pub use parser::CmXmlParser;